+++ exts/omni.isaac.lab_quadruped_tasks/omni/isaac/lab_quadruped_tasks/__init__.py
@@ -1,6 +1,7 @@
 """
-Python module serving as a project/extension template.
+License: MIT License
+Copyright (c) 2024, Felipe Mohr Santos
 """
 
-# Register Gym environments.
-from . import quadruped_env_cfg
+from . import quadruped_env_cfg  # noqa: F401, F403
+from . import unitree_go2  # noqa: F401, F403