+++ exts/omni.isaac.lab_quadruped_tasks/omni/isaac/lab_quadruped_tasks/mdp/observations.py
@@ -1,0 +1,25 @@
+"""
+License: MIT License
+Copyright (c) 2024, Felipe Mohr Santos
+"""
+
+from __future__ import annotations
+
+import torch
+from typing import TYPE_CHECKING
+
+from omni.isaac.lab.managers import SceneEntityCfg
+from omni.isaac.lab.sensors import ContactSensor
+
+if TYPE_CHECKING:
+    from omni.isaac.lab.envs import ManagerBasedRLEnv
+
+
+def feet_contact_bools(env: ManagerBasedRLEnv, sensor_cfg: SceneEntityCfg, threshold: float) -> torch.Tensor:
+    """Feet contact booleans. The foot is in contact when the force sensor exceeds the threshold"""
+
+    # extract the used quantities (to enable type-hinting)
+    contact_sensor: ContactSensor = env.scene.sensors[sensor_cfg.name]
+    net_contact_forces = contact_sensor.data.net_forces_w
+    # check which contact forces exceed the threshold
+    return torch.norm(net_contact_forces[:, sensor_cfg.body_ids], dim=-1) > threshold