+++ exts/omni.isaac.lab_quadruped_tasks/docs/CHANGELOG.rst
@@ -1,10 +1,12 @@
 Changelog
 ---------
 
-0.1.0 (2024-01-29)
+0.1.0 (2024-07-03)
 ~~~~~~~~~~~~~~~~~~
 
 Added
 ^^^^^
 
-* Created an initial template for building an extension or project based on Isaac Lab
+* Created Unitree Go2 RSL-RL agent
+* Created `QuadrupedEnvCfg` environment
+* Created `omni.isaac.lab_quadruped_tasks` extension for Isaac Lab